LayoutTests/fast/text/trailing-word-detection.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test-pre.js"></script>
<style>
@supports (-apple-trailing-word: auto) {
    #test {
        color: rgb(0, 128, 0);
    }
}
</style>
</head>
<body>
<div id="test"></div>
<script>
let element = document.getElementById("test");
shouldBeTrue("'-apple-trailing-word' in element.style");
shouldBeEqualToString("window.getComputedStyle(element).getPropertyValue('color')", "rgb(0, 128, 0)");
</script>
<script src="../../resources/js-test-post.js"></script>
</body>
</html>
