LayoutTests/fast/text/trailing-word-detection-expected.txt

PASS '-apple-trailing-word' in element.style is true
PASS window.getComputedStyle(element).getPropertyValue('color') is "rgb(0, 128, 0)"
PASS successfullyParsed is true

TEST COMPLETE
